notification about high scores giving coins;

+++ Penguin Rescue/LevelPackManager.h
@@ -21,7 +21,7 @@
 
 +(NSDictionary*)level:(NSString*)levelPath inPack:(NSString*)packPath;
 
-+(void)completeLevel:(NSString*)levelPath inPack:(NSString*)packPath withScore:(double)score;
++(void)completeLevel:(NSString*)levelPath inPack:(NSString*)packPath withScore:(int)score;
 +(NSNumber*)scoreForLevel:(NSString*)levelPath inPack:(NSString*)packPath;
 
 +(NSString*)levelAfter:(NSString*)levelPath inPack:(NSString*)packPath;
